BackEnd/insight/requirements.txt
django
djangorestframework
django-cors-headers==3.13.0
youtube_transcript_api
google-api-core
google-auth
google-cloud-language
google-cloud-speech
googleapis-common-protos
grpcio
psycopg2-binary
django-redis

